requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.35
psycopg[binary]==3.3.3
python-dotenv==1.0.1